+++ src/components/20-molecules/datepicker/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 7.0.5
 
 - Fix wrong date formatting for locale `it-CH` (#1740)
+- Now all month identifiers of all languages are capitalized at IE browser.
 
 ## 7.0.4